--- src/_main/VTaskDashboard/components/VTaskDash.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const VTaskDash = () => {
-  return <div>VTaskDash</div>;
-};
